backend/pyproject.toml
[tool.poetry]
name = "compass backend"
version = "0.0.0"
description = ""
authors = ["Tabiya <tabiya@tabiya.org>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = ">=0.23.2,<0.24.0"
python-dotenv = "^1.0.1"
deprecated="^1.2.14"
google-cloud-dlp="^3.16.0"
fix-busted-json=">=0.0.17,<1.0.0"
json-repair=">=0.44.1,<1.0.0"
motor = "^3.4.0"
google-cloud-aiplatform = "^1.53.0"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.35.0"}
pympler="^1.1"

scipy = "^1.16.1"
markitdown = {extras = ["pdf", "docx"], version = "^0.1.2"}
brotli = "^1.1.0"
httpx = "^0.28.1"
openpyxl = "^3.1.5"
selenium = "^4.38.0"
webdriver-manager = "^4.0.2"
thefuzz = "^0.22.1"
python-levenshtein = "^0.27.3"
[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-mock = "^3.14.0"
tqdm="^4.66"
bandit = "^1.7.8"
pylint = "^3.1.0"
pylint-pydantic="^0.3.2"
datasets = "^2.19.2"
google-cloud = "^0.34.0"
firebase-admin = "6.5.0"
pymongo_inmemory = "^0.4.2"
pytest-repeat = "^0.9.4"
matplotlib="^3.9.0"
google-genai="^1.26.0"
google-cloud-storage = "^2.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
export_openapi = "scripts.export_openapi:export"
convert_openapi = "scripts.convert_to_openapi2:convert"
